integration field is now float based, making it more percise

+++ Assets/Scripts/Path Finding/Integration Field/IntFieldResetJob.cs
@@ -16,7 +16,7 @@ public struct IntFieldResetJob : IJobParallelFor
     public IntFieldResetJob(NativeArray<IntegrationTile> integrationField)
     {
         IntegrationField = integrationField;
-        resetTile = new IntegrationTile(int.MaxValue, IntegrationMark.Absolute);
+        resetTile = new IntegrationTile(float.MaxValue, IntegrationMark.Absolute);
     }
     public void Execute(int index)
     {
